requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==3.2.13
django-allauth==0.51.0
django-choices==1.7.2
django-ckeditor==6.4.2
django-cors-headers==3.13.0
django-filter==22.1
django-js-asset==2.0.0
django-phonenumber-field==6.3.0
djangorestframework==3.13.1
idna==3.3
Markdown==3.3.7
oauthlib==3.2.0
Pillow==9.1.1
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.28.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
